+++ src/app/em-annotation-summary/em-annotation-summary.module.ts
@@ -1,13 +1,10 @@
 import {APP_INITIALIZER, NgModule} from '@angular/core';
 import {HttpClientModule} from '@angular/common/http';
 import {BrowserModule} from '@angular/platform-browser';
-import {FormsModule} from '@angular/forms';
 import {CookieModule} from 'ngx-cookie';
 import {AppConfig} from '../app.config';
-import {SessionService} from '../auth/session.service';
 import {EmAnnotationSummaryComponent} from '../em-annotation-summary/em-annotation-summary.component';
-import {AnnotationService} from '../dm-viewer/annotations/annotation.service';
-import {WindowService} from '../utils/window.service';
+import {AnnotationService} from '../em-viewer/annotations/annotation.service';
 
 @NgModule({
   declarations: [
@@ -20,8 +17,6 @@ import {WindowService} from '../utils/window.service';
     CookieModule.forRoot()
   ],
   providers: [
-    SessionService,
-    WindowService,
     AnnotationService,
     AppConfig,
     {
