+++ src/app/em-viewer/annotations/annotation.service.ts
@@ -1,7 +1,6 @@
 import {Injectable} from '@angular/core';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
 import {AppConfig} from '../../app.config';
-import {SessionService} from '../../auth/session.service';
 import {Observable} from 'rxjs/Observable';
 
 const pageNoteType = 'PAGENOTE';
