+++ src/app/em-viewer/annotations/notes/notes.component.ts
