+++ src/app/em-viewer/em-viewer.module.ts
@@ -1,30 +1,22 @@
 import {APP_INITIALIZER, NgModule} from '@angular/core';
-import {DmViewerComponent} from './dm-viewer.component';
+import {EmViewerComponent} from './em-viewer.component';
 import {HttpClientModule} from '@angular/common/http';
 import {BrowserModule} from '@angular/platform-browser';
 import {PdfViewerModule} from 'ng2-pdf-viewer';
 import {FormsModule} from '@angular/forms';
 import {CookieModule} from 'ngx-cookie';
-import {WindowService} from '../utils/window.service';
 import {AppConfig} from '../app.config';
-import {IdamGuard} from '../auth/idam.guard';
-import {DocumentService} from '../utils/document.service';
-import {SessionService} from '../auth/session.service';
 import {ViewerFactoryService} from './viewers/viewer-factory.service';
 import {AnnotationService} from './annotations/annotation.service';
-import {DmViewerRouteComponent} from './dm-viewer-route.component';
 import {PdfViewerComponent} from './viewers/pdf-viewer/pdf-viewer.component';
 import {UnsupportedViewerComponent} from './viewers/unsupported-viewer/unsupported-viewer.component';
-import {ImagePipe} from '../utils/image-pipe';
 import {NotesComponent} from './annotations/notes/notes.component';
-import {EmAnnotationSummaryComponent} from '../em-annotation-summary/em-annotation-summary.component';
-import {EmAnnotationSummaryRouteComponent} from '../em-annotation-summary/em-annotation-summary-route.component';
 import {ImgViewerComponent} from './viewers/img-viewer/img-viewer.component';
 import {ViewerAnchorDirective} from './viewers/viewer-anchor.directive';
 
 @NgModule({
   declarations: [
-    DmViewerComponent,
+    EmViewerComponent,
     PdfViewerComponent,
     ImgViewerComponent,
     UnsupportedViewerComponent,
@@ -36,7 +28,7 @@ import {ViewerAnchorDirective} from './viewers/viewer-anchor.directive';
     ImgViewerComponent,
     UnsupportedViewerComponent
   ],
-  exports: [DmViewerComponent],
+  exports: [EmViewerComponent],
   imports: [
     BrowserModule,
     FormsModule,
