+++ src/app/em-viewer/viewers/img-viewer/img-viewer.component.scss
