+++ src/app/em-viewer/viewers/pdf-viewer/pdf-viewer.component.scss
