+++ src/app/em-viewer/viewers/viewer-anchor.directive.ts
