--- src/app/utils/document.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable()
-export class DocumentService {
-  getUrl(): string {
-    return document.URL;
-  }
-}
