--- src/app/utils/window.service.ts
@@ -1,12 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable()
-export class WindowService {
-  locationAssign(url: string): void {
-    window.location.assign(url);
-  }
-
-  reload() {
-    window.location.reload();
-  }
-}
